requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
ollama>=0.2.1
openai>=1.30.0
apscheduler>=3.10.4
python-telegram-bot>=21.3
pyyaml>=6.0.1
aiosqlite>=0.20.0
websockets>=12.0
